+++ apps/web/src/lib/icons/index.ts
@@ -1,3 +1,3 @@
-export { default as GitHub } from './git-hub.svelte';
+export { default as GitHub } from './github.svelte';
 export { default as Twitter } from './twitter.svelte';
-export { default as YouTube } from './you-tube.svelte';
+export { default as YouTube } from './youtube.svelte';
