+++ apps/web/src/lib/icons/youtube.svelte
